move lib functions to core package
