+++ packages/cli/src/commands/account/create.ts
@@ -2,10 +2,10 @@ import { Command, Flags } from "@oclif/core";
 import chalk = require("chalk");
 import { writeJSON } from "fs-extra";
 import { ChainAccount } from "../../lib/account";
-import { ensureSwankyProject, getSwankyConfig } from "../../lib/command-utils";
-import { encrypt } from "../../lib/crypto";
+import { ensureSwankyProject, getSwankyConfig } from "@astar-network/swanky-core";
+import { encrypt } from "@astar-network/swanky-core";
 import inquirer from "inquirer";
-import { AccountData } from "../../types";
+import { AccountData } from "@astar-network/swanky-core";
 export class CreateAccount extends Command {
   static description = "Create a new dev account in config";
