+++ packages/cli/src/commands/account/list.ts
@@ -1,6 +1,6 @@
 import { Command } from "@oclif/core";
 import chalk = require("chalk");
-import { ensureSwankyProject, getSwankyConfig } from "../../lib/command-utils";
+import { AccountData, ensureSwankyProject, getSwankyConfig } from "@astar-network/swanky-core";
 
 export class CreateAccount extends Command {
   static description = "List dev accounts stored in config";
@@ -12,7 +12,7 @@ export class CreateAccount extends Command {
     const config = await getSwankyConfig();
     this.log(`${chalk.greenBright("✔")} Stored dev accounts:`);
 
-    config.accounts.forEach((account: any) => {
+    config.accounts.forEach((account: AccountData) => {
       this.log(`\t${chalk.yellowBright("Alias: ")} ${account.alias}`);
     });
   }
