+++ packages/cli/src/commands/node/purge.ts
@@ -1,6 +1,6 @@
 import { Command } from "@oclif/core";
 import execa from "execa";
-import { ensureSwankyProject, getSwankyConfig } from "../../lib/command-utils";
+import { ensureSwankyProject, getSwankyConfig } from "@astar-network/swanky-core";
 export class PurgeNode extends Command {
   static description = "Purge local chain state";
