+++ packages/core/package.json
@@ -4,10 +4,16 @@
   "description": "Core library of Swanky Suite",
   "author": "Astar network",
   "license": "MIT",
+  "main": "./dist/index",
   "repository": "https://github.com/AstarNetwork/swanky-cli",
-  "dependencies": {},
+  "dependencies": {
+    "execa": "5.1.1",
+    "fs-extra": "^10.1.0"
+  },
   "devDependencies": {},
-  "scripts": {},
+  "scripts": {
+    "build": "shx rm -rf dist && rm -f tsconfig.tsbuildinfo && tsc -b"
+  },
   "engines": {
     "node": ">=18.0.0"
   }
