+++ packages/core/src/types/index.ts
@@ -1,4 +1,4 @@
-import { SubmittableExtrinsic } from "@polkadot/api/types";
+// import { SubmittableExtrinsic } from "@polkadot/api/types";
 
 export type KeypairType = "ed25519" | "sr25519" | "ecdsa" | "ethereum";
 
@@ -9,7 +9,7 @@ export interface ChainProperty {
   ss58Prefix: number;
 }
 
-export type ExtrinsicPayload = SubmittableExtrinsic<"promise">;
+// export type ExtrinsicPayload = SubmittableExtrinsic<"promise">;
 
 export type Encrypted = { iv: string; data: string };
